Modified spectrum generator to use constraint handler class
This is probably a temporary solution for this project, as
the solvers are similar enough that this works (and avoids
having to write a separate spectrum generator class). In the
long term it will probably be better (e.g. for more diverse
solvers) to have a separate specialisation in each case.